Firmware v1: prod_v1_pro board header from real hardware, and a release build
prod_v1_pro.h: filled in from the shipped Y11 netlist and CONFIRMED on board #1
during first-article bring-up. LED_DATA=10 (NOT GPIO2 - stock WLED's C3 default
is a strap tied to +3V3 here), BUTTON_PRIMARY=9 shared with the boot strap,
BUTTON_RESET=-1 because SW1 is wired to hardware EN and firmware cannot read it,
LED_POWER_EN=-1 (no rail gate on rev 1 - a rev-2 candidate). Plus I2C/VBUS-sense/
PG extras, the reserved-pin list, and a note that GPIO0/1 are the two free pins.
The #error is removed; the target builds and runs.

Release build: the project was compiling at -Og with full assertions, INFO
logging, the esp_err_t name table and the FULL Mozilla root CA bundle - the
bundle alone measured 118 KB, not the '~64KB compressed' its own comment claimed.
Five settings, no features removed: 1,297,088 -> 1,036,160 B, 255 KB / 20% off,
headroom 1% -> 21%. Assertions are SILENT not DISABLE, deliberately: with
BOOTLOADER_APP_ROLLBACK_ENABLE we want a bad image to abort and roll back.

CONFIG_IDF_TARGET pinned in sdkconfig.defaults. It previously lived ONLY in the
generated sdkconfig, so deleting that file silently retargeted the whole project
to classic ESP32 - it built clean and reported a BETTER size, with a 0x1000
bootloader offset instead of the C3's 0x0.

ll_wifi.c: %hhu instead of %u in entry_key. Not a bug - idx is uint8_t so the
worst case is 10 bytes into a 16-byte buffer - but the (unsigned) cast discards
the range, and at -Os GCC inlines three frames deep, assumes 32 bits, computes
17 bytes and trips -Werror=format-truncation.

+++ Firmware/v1/core/ll_wifi/ll_wifi.c
@@ -34,7 +34,12 @@ static bool             g_inited;
 
 static void entry_key(uint8_t idx, char *out, size_t out_size)
 {
-    snprintf(out, out_size, "entry_%u", (unsigned)idx);
+    /* %hhu, not %u: idx is a uint8_t so the output is at most "entry_255"
+     * (10 bytes incl. NUL) into a 16-byte buffer. The (unsigned) cast threw
+     * that range away, and at -Os GCC inlines entry_key -> erase_entry ->
+     * persist_locked, assumes the full 32-bit range, computes 17 bytes and
+     * trips -Werror=format-truncation. No behaviour change. 2026-08-25. */
+    snprintf(out, out_size, "entry_%hhu", idx);
 }
 
 static esp_err_t open_rw(nvs_handle_t *h)
